requirements.txt
azure-cognitiveservices-vision-computervision
openai>=1.0.0
